% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antiTNFresponse.R
\name{antiTNFresponse}
\alias{antiTNFresponse}
\title{ATRPred: Anti-TNF Treatment Response Predictor}
\usage{
antiTNFresponse()
}
\description{
This function predicts the response to anti-TNF therapy in Rheumatoid Arthritis patients at baseline
It requires an excel input for 17 proteins along with gender and baseline DAS
Any values not present will be imputed
Please don't change the order of proteins in the sample excel file
}
